chore: refactor `cast` for library use (#10123)
