+++ crates/cast/Cargo.toml
@@ -98,7 +98,3 @@ asm-keccak = ["alloy-primitives/asm-keccak"]
 jemalloc = ["dep:tikv-jemallocator"]
 aws-kms = ["foundry-wallets/aws-kms", "dep:aws-sdk-kms"]
 isolate-by-default = ["foundry-config/isolate-by-default"]
-
-[[bench]]
-name = "vanity"
-harness = false
