+++ crates/cast/src/cmd/logs.rs
@@ -1,9 +1,9 @@
+use crate::Cast;
 use alloy_dyn_abi::{DynSolType, DynSolValue, Specifier};
 use alloy_json_abi::Event;
 use alloy_network::AnyNetwork;
 use alloy_primitives::{hex::FromHex, Address, B256};
 use alloy_rpc_types::{BlockId, BlockNumberOrTag, Filter, FilterBlockOption, FilterSet, Topic};
-use cast::Cast;
 use clap::Parser;
 use eyre::Result;
 use foundry_cli::{opts::EthereumOpts, utils, utils::LoadConfig};
