+++ crates/cast/src/opts.rs
@@ -528,7 +528,7 @@ pub enum CastSubcommand {
     /// Decode event data.
     #[command(visible_aliases = &["event-decode", "--event-decode", "ed"])]
     DecodeEvent {
-        /// The event signature. If none provided then tries to decode from local cache or `https://api.openchain.xyz`.
+        /// The event signature. If none provided then tries to decode from local cache or <https://api.openchain.xyz>.
         #[arg(long, visible_alias = "event-sig")]
         sig: Option<String>,
         /// The event data to decode.
@@ -538,7 +538,7 @@ pub enum CastSubcommand {
     /// Decode custom error data.
     #[command(visible_aliases = &["error-decode", "--error-decode", "erd"])]
     DecodeError {
-        /// The error signature. If none provided then tries to decode from local cache or `https://api.openchain.xyz`.
+        /// The error signature. If none provided then tries to decode from local cache or <https://api.openchain.xyz>.
         #[arg(long, visible_alias = "error-sig")]
         sig: Option<String>,
         /// The error data to decode.
@@ -642,21 +642,21 @@ pub enum CastSubcommand {
         rpc: RpcOpts,
     },
 
-    /// Get the function signatures for the given selector from https://openchain.xyz.
+    /// Get the function signatures for the given selector from <https://openchain.xyz>.
     #[command(name = "4byte", visible_aliases = &["4", "4b"])]
     FourByte {
         /// The function selector.
         selector: Option<String>,
     },
 
-    /// Decode ABI-encoded calldata using https://openchain.xyz.
+    /// Decode ABI-encoded calldata using <https://openchain.xyz>.
     #[command(name = "4byte-calldata", aliases = &["4byte-decode", "4d", "4bd"], visible_aliases = &["4c", "4bc"])]
     FourByteCalldata {
         /// The ABI-encoded calldata.
         calldata: Option<String>,
     },
 
-    /// Get the event signature for a given topic 0 from https://openchain.xyz.
+    /// Get the event signature for a given topic 0 from <https://openchain.xyz>.
     #[command(name = "4byte-event", visible_aliases = &["4e", "4be", "topic0-event", "t0e"])]
     FourByteEvent {
         /// Topic 0
@@ -664,7 +664,7 @@ pub enum CastSubcommand {
         topic: Option<String>,
     },
 
-    /// Upload the given signatures to https://openchain.xyz.
+    /// Upload the given signatures to <https://openchain.xyz>.
     ///
     /// Example inputs:
     /// - "transfer(address,uint256)"
@@ -682,13 +682,13 @@ pub enum CastSubcommand {
 
     /// Pretty print calldata.
     ///
-    /// Tries to decode the calldata using https://openchain.xyz unless --offline is passed.
+    /// Tries to decode the calldata using <https://openchain.xyz> unless --offline is passed.
     #[command(visible_alias = "pc")]
     PrettyCalldata {
         /// The calldata.
         calldata: Option<String>,
 
-        /// Skip the https://openchain.xyz lookup.
+        /// Skip the <https://openchain.xyz> lookup.
         #[arg(long, short)]
         offline: bool,
     },
@@ -1044,7 +1044,7 @@ pub enum CastSubcommand {
         /// The hex-encoded bytecode.
         bytecode: Option<String>,
 
-        /// Resolve the function signatures for the extracted selectors using https://openchain.xyz
+        /// Resolve the function signatures for the extracted selectors using <https://openchain.xyz>
         #[arg(long, short)]
         resolve: bool,
     },
@@ -1074,8 +1074,8 @@ pub fn parse_slot(s: &str) -> Result<B256> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::SimpleCast;
     use alloy_rpc_types::{BlockNumberOrTag, RpcBlockHash};
-    use cast::SimpleCast;
     use clap::CommandFactory;
 
     #[test]
